+++ public/index.html
@@ -4,11 +4,11 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Jobs List</title>
+    <title>Plants List</title>
   </head>
 
   <body>
-    <h1>Jobs List</h1>
+    <h1>Plants List</h1>
     <hr />
     <p id="message"></p>
     <hr />
@@ -52,8 +52,8 @@
         <button type="button" id="register-cancel">cancel</button>
       </form>
     </div>
-    <div id="jobs" style="display:none;">
-      <p id="jobs-message"></p>
+    <div id="items" style="display:none;">
+      <p id="items-message"></p>
       <table id="items-table">
         <tr id="items-table-header">
           <th>Name</th>
